models and db models created for handling db storage and persistency
